backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
celery==5.3.4
redis==5.0.1
python-decouple==3.8
django-extensions==3.2.3
gunicorn==21.2.0

# RAG Pipeline Dependencies
langchain==0.0.350
langchain-community==0.0.3
langchain-huggingface==0.0.1
qdrant-client==1.7.0
transformers==4.35.2
torch==2.1.1
pdfplumber==0.9.0
newspaper3k==0.2.8
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.3

# Additional utilities
Pillow==10.1.0
python-magic==0.4.27
celery-progress==0.3